Implement URLReferencedModelListField
A common pattern in the API is that the user can add references to other
relevant objects using a list of identifying URLs. However,
Metax would not validate user input for these reference fields properly,
with nonexistent URLs or malformed input causing an unhandled
exception instead of an user readable error. In addition, each single
object was retrieved in a separate query, resulting in unnecessary DB
load.

The `URLReferencedModelListField` serializer field takes care of
validating the user input for these reference fields and generating
per-field error messages if necessary. It also retrieves the objects in
bulk.

Refs CSCMETAXREWORK-116