Fix PAS compatible file copies
When a PAS catalog copy of a dataset is created, we also have to update
the references to (non) PAS compatible files or the operation will fail.

The previous commit ensures that the PAS file pairs in the dataset are
complete (i.e. for every dataset file with a linking both PAS and non
PAS compatible files exist in the dataset).

Refs TPASPKT-1378