Validate PAS compatible file pairs
Validate PAS compatible file pairs when creating a PAS catalog copy of a
dataset. This means that if a dataset contains a file with PAS
compatible/non PAS compatible counterpart, then both files have to be
included with the same dataset.

At the moment this validation is only done when creating a PAS copy of a
dataset. The validation method could be called in other contexts if this
is deemed necessary later, however.

Refs TPASPKT-1378